feat: use typescript
