+++ .gitignore
@@ -1,2 +1,3 @@
+dist
 *.log
 node_modules
