+++ package.json
@@ -2,19 +2,53 @@
   "name": "wamp-cli",
   "version": "1.0.2",
   "description": "A CLI for testing WAMP connections",
-  "main": "index.js",
+  "main": "dist/src/repl.js",
   "bin": {
-    "wampc": "bin/wampc"
+    "wampc": "dist/src/bin/program.js"
   },
   "author": "John Wright <johngeorge.wright@gmail.com>",
   "license": "MIT",
+  "scripts": {
+    "prebuild": "npm run clean",
+    "build": "tsc",
+    "postbuild": "rimraf 'dist/**/?(__tests__|__mocks__|__setup__|*.test.*)'",
+    "clean": "rimraf dist",
+    "start": "ts-node src/program"
+  },
   "dependencies": {
     "autobahn": "^20.7.2",
     "cli-table": "^0.3.1",
-    "colors": "^1.4.0",
-    "commander": "^2.20.3"
+    "colors": "^1.4.0"
   },
   "devDependencies": {
-    "eslint": "^7.8.1"
+    "@commitlint/cli": "^9.1.2",
+    "@commitlint/config-conventional": "^9.1.2",
+    "@types/autobahn": "^18.10.0",
+    "@types/cli-table": "^0.3.0",
+    "@types/node": "^14.6.4",
+    "commander": "^6.1.0",
+    "cz-conventional-changelog": "^3.3.0",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.3.0",
+    "prettier": "^2.1.1",
+    "rimraf": "^3.0.2",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.2"
+  },
+  "husky": {
+    "hooks": {
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.{md,json,js,jsx,ts,tsx,yml,yaml}": [
+      "prettier --ignore-path .gitignore --write"
+    ]
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
   }
 }
