+++ prettier.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  semi: false,
+  singleQuote: true,
+}
